+++ api/github.ts
@@ -20,8 +20,8 @@ export const getCategories = async () => {
  *
  * @return ['토비', '제로초', ...]
  */
-export const getCreator = async (path: string) => {
-  const res = await fetch(`${API_URL}/${path}`, {
+export const getCreator = async (category: string) => {
+  const res = await fetch(`${API_URL}/${category}`, {
     next: { revalidate: REVALIDATE_GITHUB_DATA },
   });
   const data: GithubMarkdown[] = await res.json();
@@ -34,4 +34,25 @@ export const getCreator = async (path: string) => {
  *
  * @return ['토비-스프링-부트-이해와-원리', 'Slack-클론-코딩-실시간-채팅-with-React', ...]
  */
-export const getLecture = async () => {};
+export const getLecture = async (category: string) => {
+  const res = await fetch(`${API_URL}/${category}`, {
+    next: { revalidate: REVALIDATE_GITHUB_DATA },
+  });
+  const creators: GithubMarkdown[] = await res.json();
+
+  let lecturesAll = [];
+
+  for (const { path } of creators) {
+    const lectureByCreator: GithubMarkdown[] = await fetch(`${API_URL}/${path}`, {
+      next: { revalidate: REVALIDATE_GITHUB_DATA },
+    });
+
+    for (const lectureInfo of lectureByCreator) {
+      if (lectureInfo.name !== 'image') {
+        lecturesAll.push(lectureInfo);
+      }
+    }
+  }
+
+  return lecturesAll;
+};
